# Inlining for easy performance win