--- a/target/cheri-common/op_helper_cheri_common.c
+++ b/target/cheri-common/op_helper_cheri_common.c
@@ -1013,60 +1013,9 @@ const cap_register_t *get_load_store_base_cap(CPUArchState *env, uint32_t cb)
 #endif
 }
 
-target_ulong cap_check_common_reg(uint32_t required_perms, CPUArchState *env,
-                                  uint32_t cb, target_ulong offset,
-                                  uint32_t size, uintptr_t _host_return_address,
-                                  const cap_register_t *cbp,
-                                  uint32_t alignment_required,
-                                  unaligned_memaccess_handler unaligned_handler)
-{
-#define MISSING_REQUIRED_PERM(X) ((required_perms & ~cbp->cr_perms) & (X))
-    if (!cbp->cr_tag) {
-        raise_cheri_exception(env, CapEx_TagViolation, cb);
-    } else if (is_cap_sealed(cbp)) {
-        raise_cheri_exception(env, CapEx_SealViolation, cb);
-    } else if (MISSING_REQUIRED_PERM(CAP_PERM_LOAD)) {
-        raise_cheri_exception(env, CapEx_PermitLoadViolation, cb);
-    } else if (MISSING_REQUIRED_PERM(CAP_PERM_STORE)) {
-        raise_cheri_exception(env, CapEx_PermitStoreViolation, cb);
-    } else if (MISSING_REQUIRED_PERM(CAP_PERM_STORE_CAP)) {
-        raise_cheri_exception(env, CapEx_PermitStoreCapViolation, cb);
-    } else if (MISSING_REQUIRED_PERM(CAP_PERM_STORE_LOCAL)) {
-        raise_cheri_exception(env, CapEx_PermitStoreLocalCapViolation, cb);
-    }
-#undef MISSING_REQUIRED_PERM
-
-    const target_ulong cursor = cap_get_cursor(cbp);
-    const target_ulong addr = cursor + (target_long)offset;
-    if (!cap_is_in_bounds(cbp, addr, size)) {
-        qemu_log_instr_or_mask_msg(env, CPU_LOG_INT,
-            "Failed capability bounds check: offset=" TARGET_FMT_lx
-            " cursor=" TARGET_FMT_lx " addr=" TARGET_FMT_lx "\n",
-            offset, cursor, addr);
-        raise_cheri_exception(env, CapEx_LengthViolation, cb);
-    } else if (alignment_required &&
-               !QEMU_IS_ALIGNED_P2(addr, alignment_required)) {
-        if (unaligned_handler) {
-            unaligned_handler(env, addr, _host_return_address);
-        }
-#if defined(TARGET_MIPS) && defined(CHERI_UNALIGNED)
-        const char *access_type =
-            (required_perms == (CAP_PERM_STORE | CAP_PERM_LOAD))
-                ? "RMW"
-                : ((required_perms == CAP_PERM_STORE) ? "store" : "load");
-        qemu_maybe_log_instr_extra(env,
-                                   "Allowing unaligned %d-byte %s of "
-                                   "address 0x%" PRIx64 "\n",
-                                   size, access_type, addr);
-#endif
-    }
-    return addr;
-}
-
-static inline target_ulong cap_check_common(uint32_t required_perms,
-                                            CPUArchState *env, uint32_t cb,
-                                            target_ulong offset, uint32_t size,
-                                            uintptr_t _host_return_address)
+static inline QEMU_ALWAYS_INLINE target_ulong cap_check_common(
+    uint32_t required_perms, CPUArchState *env, uint32_t cb,
+    target_ulong offset, uint32_t size, uintptr_t _host_return_address)
 {
     const cap_register_t *cbp = get_load_store_base_cap(env, cb);
     return cap_check_common_reg(required_perms, env, cb, offset, size,